% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_and_replace.R
\name{find_and_replace}
\alias{find_and_replace}
\title{Find and replace all variables in one of the files of the book skeleton}
\usage{
find_and_replace(
  file,
  replacements = c(book_title = "My book", author_name = "Marion Louveaux",
    creation_date = "`r Sys.Date()`", short_description =
    "An example of book with {bookprep}.", index_title = "Context")
)
}
\arguments{
\item{file}{path of a file from skeleton: _bookdown.yml, _output.yml, index.Rmd, README.md}

\item{replacements}{a vector of correspondences between an input and an output. Input is fixed and corresponds to: book_title, author_name, creation_date, short_description and index_title}
}
\value{
path of the modified file
}
\description{
Find and replace all variables in one of the files of the book skeleton
}
\examples{
# copy to a temporary file to avoid overwriting the package
index_tmp <- tempfile(pattern = "index-")

file.copy(from = system.file("book_skeleton/index.Rmd", package = "bookprep"), to = index_tmp)

bookprep:::find_and_replace(
  file = index_tmp,
  replacements = c(
    "book_title" = "My book",
    "author_name" = "Marion Louveaux",
    "creation_date" = "`r Sys.Date()`",
    "short_description" = "An example of book with {bookprep}.",
    "index_title" = "Context"
  )
)
}
