% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_template.R
\name{initialize_template}
\alias{initialize_template}
\title{Initialize a book template with named chapters and base content}
\usage{
initialize_template(
  path,
  chapters = c("Preface", "Introduction"),
  references = "References"
)
}
\arguments{
\item{path}{where to create the template}

\item{chapters}{titles of Rmd files to include in the template}

\item{references}{title of the references Rmd or NULL}
}
\value{
used for side effect: file creation in path
}
\description{
Initialize a book template with named chapters and base content
}
\examples{
# Create temporary directory for reproducible example
dir_tmp <- tempfile(pattern = "proj-")
dir.create(dir_tmp)
# browseURL(dir_tmp)

initialize_template(
  path = dir_tmp,
  chapters = c(
    "Introduction", "Material and Methods",
    "Results", "Discussion"
  ),
  references = NULL
)
}
