% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_files.R
\name{prepare_files}
\alias{prepare_files}
\title{Modify all variables in files from the book skeleton}
\usage{
prepare_files(
  path,
  replacements = c(book_title = "My book", author_name = "Marion Louveaux",
    creation_date = "`r Sys.Date()`", short_description =
    "An example of book with {bookprep}.", index_title = "Context"),
  pattern = "[.]Rmd$|[.]yml$|[.]md$"
)
}
\arguments{
\item{path}{path to directory (skeleton) containing all files to modify}

\item{replacements}{a vector of correspondences between an input and an output. Input is fixed and corresponds to: book_title, author_name, creation_date, short_description and index_title}

\item{pattern}{regex to identify files to prepare (by default, all Rmd, yml and md files)}
}
\value{
path from each modified file
}
\description{
Modify all variables in files from the book skeleton
}
\examples{
# Create temporary directory for reproducible example
dir_tmp <- tempfile(pattern = "proj-")
dir.create(dir_tmp)
# browseURL(dir_tmp)

# prepare_files() requires a directory with some files inside to look for replacements
file.copy(from = system.file("book_skeleton", package = "bookprep"), to = dir_tmp, recursive = TRUE)

bookprep:::prepare_files(
  path = dir_tmp,
  replacements = c(
    "book_title" = "My book",
    "author_name" = "Marion Louveaux",
    "creation_date" = "`r Sys.Date()`",
    "short_description" = "An example of book with {bookprep}.",
    "index_title" = "Context"
  )
)
}
